lsu_pilot/pyproject.toml
[tool.poetry]
name = "lsu-pilot"
version = "0.1.0"
description = ""
authors = ["John Liu <stoic.llama@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.7"
python-telegram-bot = "^21.1.1"
openai = "1.13.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
tiktoken = "^0.6.0"
langchain = "^0.1.17"
numpy = "^1.26.4"
scipy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"